Recovered changes that were destroyed previously
Some of the code is derived from a decompiled jar. I have done my best to make readable some of the parts that are totally unreadable. To be clear, only Helper.java and Commands.java contains any code deriving from a decompiled jar. Anyway, I am putting all of this on this branch because while I think it will work, we have to test it in a multiuser environment.